telemetry/observability-signal-ingestor/requirements.txt
requests==2.32.3
kubernetes==29.0.0
pyyaml==6.0.2
